+++ toro-sample/src/main/java/im/ene/lab/toro/sample/presentation/average1/Average1NormalViewHolder.java
@@ -14,23 +14,26 @@
  * limitations under the License.
  */
 
-package im.ene.lab.toro.sample.average1;
+package im.ene.lab.toro.sample.presentation.average1;
 
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.RecyclerView;
+import android.view.View;
+import im.ene.lab.toro.sample.R;
 
 /**
- * Created by eneim on 6/30/16.
+ * Created by eneim on 6/29/16.
+ *
+ * Sample ViewHolder which holds no Video. Or it doesn't request support from Toro.
  */
-public class Average1Activity extends AppCompatActivity {
+public class Average1NormalViewHolder extends Average1ViewHolder {
 
-  @Override protected void onCreate(@Nullable Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-    if (savedInstanceState == null) {
-      getSupportFragmentManager().beginTransaction()
-          .replace(android.R.id.content, Average1ListFragment.newInstance())
-          .commit();
-    }
+  public static final int LAYOUT_RES = R.layout.vh_normal_view;
+
+  public Average1NormalViewHolder(View itemView) {
+    super(itemView);
+  }
+
+  @Override public void bind(RecyclerView.Adapter adapter, Object item) {
+
   }
 }
